flatten the pureport_client package
This commit flattens the package and simplifies the interface to remove
complexity from the implementation.  Where possible, code paths were not
changed; however, some of the module names and code has change to
support simplifying the implementation.
