+++ tox.ini
@@ -1,21 +1,52 @@
+# tox (https://tox.readthedocs.io/) is a tool for running test
+# in multiple virtualenvs. This configuration file will run the
+# test suite on all supported python versions. To use it, "pip install tox"
+# and then run "tox" from this directory.
+
+# ###########################################################################
+#
+# NOTE:
+# Tox is used as locally for executing the test cases using a matrix of 
+# different Python environments.  When code changes are pushed to the
+# main repo via pull request, the test cases are iterated using Github
+# Actions which DOES NOT use Tox.  It is import than any changes to how
+# Tox runs tests should be replicated in .github/workflows/test.yaml
+#
+# ###########################################################################
+
+
 [tox]
-skipsdist = True
-envlist = unittest,flake8
+envlist = clean, flake8, py35, py36, py37, py38, report
 
 [testenv]
 deps =
-    -rrequirements.txt
-    -rtest-requirements.txt
+    -r requirements.txt
+    -r test-requirements.txt
 
-[testenv:unittest]
-passenv = *
+setenv =
+    PYTHONDONTWRITEBYTECODE = 1
+    
+depends = 
+    {py35, py36, py37, py38}: clean
+    report: py38
+
 commands =
-    coverage run -m unittest tests/test_client.py
-    coverage html --omit="tests/**",".tox/**"
+    pytest --cov='pureport_client' --cov-append -v test/
+
+[testenv:report]
+deps = 
+    coverage 
+    flake8
+commands = 
+    coverage report -m
+    flake8 pureport_client/ tests/ --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+
+[testenv:clean]
+deps = coverage
+skip_install = true
+commands = coverage erase
 
 [testenv:flake8]
-commands =
-    flake8 pureport tests
-
-[flake8]
-max-line-length = 120
+commands = 
+    # stop the build if there are Python syntax errors or undefined names
+    flake8 pureport_client/ tests/ --count --select=E9,F63,F7,F82 --show-source --statistics
